add query types
